+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/catchup/consumer/manager/EventQueueProcessingConfig.java
@@ -1,0 +1,5 @@
+package uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager;
+
+public interface EventQueueProcessingConfig {
+    int getMaxTotalEventsInProcess();
+}